requirements.txt
# Core Django
Django>=4.2.0
djangorestframework>=3.14.0

# Database and Redis
redis>=5.0.0
psycopg2-binary>=2.9.0
django-redis>=5.4.0

# Celery
celery>=5.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# API Documentation
drf-spectacular>=0.27.0

# Environment and Utilities
python-dotenv>=1.0.0
django-filter>=23.5
django-cors-headers>=4.3.0

